telegram-bot/requirements.txt
python-telegram-bot[job-queue]>=20.7
httpx>=0.27
aiosqlite>=0.19
python-dotenv>=1.0
